requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-tailwind==1.0.3
cryptography==45.0.3
Django==5.2.1
django-allauth==65.8.1
django-countries==7.6.1
django-crispy-forms==2.4
django-unfold==0.57.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
